+++ lib/ruby_challenge/services/segment_parsers/base.rb
@@ -10,12 +10,13 @@ module RubyChallenge
     class Base
       CONFIG = {
         origin: { position: 0, type: :string },
-        from_time: { position: 1, type: :datetime },
-        to_time: { position: 2, type: :datetime }
+        from_time: { position: 1, type: :time },
+        to_time: { position: 2, type: :time }
       }.freeze
 
       def initialize(words:)
         @words = words
+        @config = CONFIG
       end
 
       def call
@@ -34,25 +35,35 @@ module RubyChallenge
           to_time: field_value(:to_time)
         )
       end
-      
+
       def check_segment_errors(segment)
         raise ParsingSegmentError, "Errors parging segment #{segment.inspect}" unless segment.valid?
       end
 
-      def field_value(field)
-        field_config = CONFIG[field]
+      def field_value(field, extra_config = {})
+        field_config = @config[field]
         position = field_config[:position]
-        field_config[:type] == :string ? @words[position] : parse_datetime(position)
+
+        case field_config[:type]
+        when :string
+          @words[position]
+        else
+          send("parse_#{field_config[:type]}", position, extra_config)
+        end
       end
 
-      def parse_datetime(position)
+      def parse_time(position, _extra_config)
         time_text = if position.is_a?(Array)
                       position.map { @words[_1] }.join(" ")
                     else
                       "#{@words[position]} 00:00"
                     end
+        convert_text_to_time(time_text)
+      end
+
+      def convert_text_to_time(time_text)
         Time.strptime(time_text, "%Y-%m-%d %H:%M")
-      rescue ArgumentError => e
+      rescue ArgumentError
         raise ParsingDateTimeError, "Invalid datetime from text '#{time_text}'"
       end
     end