+++ spec/services/segment_parsers/train_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe RubyChallenge::SegmentParsers::Train do
+  include_context "segment parsers transport tests"
+end