+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* New SegmentParsers service classes (Transport, Flight and Train) and tests
 * New SegmentParsers service classes (Base and Hotel) and tests
 * Creating patch for SvqParser and test
 * New Trip class with RSpec